Sound Bug
Found a bug in Firefox concerning the Pause Button where the loaded song seems to simply restart instead of saving the time for a playback causing everything relating to it to flickr. It could be an issue with my firefox settings. The problems are not present in Chrome,Edge and Safari.